requirements.txt
RPi.GPIO
opencv-python
numpy
picamera2
imutils
flask
flask-socketio
eventlet
